refactor(meal-form): rename directory to daily-meal-form
Renamed the 'meal-form' directory and its contents to 'daily-meal-form' for improved clarity and specificity in file organization. This change affects import paths across the codebase, including updates in 'day.tsx'. Ensure these adjustments do not impact module resolution or functionality.
