+++ src/features/daily-meal-form/components/MealIcon.tsx
